replaced weird javascript file writing system with jinja templates
