+++ strabo/strabo/static/js/drawMap-livy.js
@@ -14,7 +14,7 @@ L.tileLayer('http://{s}.tile.osm.org/{z}/{x}/{y}.png', {
 }).addTo(drawMap);
 
 // add pre-existing points, zones, and lines to map
-// interest_points, zones, and lines variables from the 
+// interest_points, zones, and lines variables from the
 // interest_points.js file
 var point_features = L.geoJson(interest_points, {
   onEachFeature: onEachPoint,
@@ -62,7 +62,7 @@ drawMap.addLayer(drawnItems);
 var shapeColorInit = '#2397EB';
 
 var options1 = {
-    
+
     draw : {
       polyline: {
         shapeOptions: {
@@ -124,7 +124,7 @@ drawMap.on('draw:created', function (e) { //grabs layer of drawn item
   }
   else if (type === 'marker') {
     drawnItems.addLayer(shapeLayer);
-    obJSON = shapeLayer.toGeoJSON();    
+    obJSON = shapeLayer.toGeoJSON();
   }
   else if (type === 'polygon') {
     obJSON = shapeLayer.toGeoJSON();
@@ -143,7 +143,7 @@ drawMap.on('draw:deletestop', function (e) {
 
 
 drawMap.on('draw:edited', function (e) {
-  
+
   var editLayers = e.layers;
   var type = e.layerType;
 
@@ -167,10 +167,10 @@ $(function()
 
   $e.change(function() {
     $usrSelect = $("#colorPick :selected").text();
-    console.log($usrSelect); 
-  
+    console.log($usrSelect);
 
-  
+
+
 
 
   //console.log($dropDown);
@@ -231,7 +231,7 @@ $('#upload-btn').click(function (e) {
   var JSONobject = JSON.stringify(obJSON);
   console.log(JSONobject);
   $('#geojson-field').attr("value", JSONobject);
-}); 
+});
 
 
   // Returns an array of the points in the path.
