--- strabo/strabo/static/js/interest_points_livy.js
@@ -1,9 +1,0 @@
-var lat_setting = 41.892695
-var long_setting = 12.495142
-var tile_src = "http://stamen-tiles-{s}.a.ssl.fastly.net/watercolor/{z}/{x}/{y}.png"
-var tile_attr1 = "Map tiles by Stamen Design, Map data by OpenStreetMap"
-var subdomains = "abcd"
-var extension = "png"
-var interest_points = {"features": [{"geometry": {"coordinates": [12.485162615776062, 41.88810153938247], "name": "Ficus Ruminalis (Ruminalis Fig)", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.710237503051758, 41.75120833715945], "name": "Mount Alba", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.48178571462631, 41.88805761040352], "name": "Ara Maxima (The Greatest Altar to Hercules)", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.50068187713623, 41.941680373919404], "name": "Atemnae", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.532482147216797, 42.01052981889534], "name": "Crustumerium", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.521324157714842, 41.98501507352482], "name": "Caenina", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.482123672962189, 41.891739552702994], "name": "Temple to Jupiter Feretrius", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.48222291469574, 41.89144804006195], "name": "Rupes Tarpea (Tarpeian Rock)", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.48858243227005, 41.890689302292934], "name": "Temple to Jupiter Stator", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.738475799560547, 42.20766785297783], "name": "Cures", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.48508483171463, 41.89235452021227], "name": "Lacus Curtius (Curtian Lake)", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.506539821624756, 41.97337057266951], "name": "Fidenae", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.48531550168991, 41.892693947795195], "name": "Temple of Janus", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.486189901828766, 41.89167166631636], "name": "Temple of Vesta", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.489365637302399, 41.88430756167117], "name": "Porta Capena (Capena Gate)", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.486484944820404, 41.89237847986553], "name": "Pila Horatia (Horatian Spears)", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.495306730270384, 41.89122441302461], "name": "Sororium Tigillium", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.4851, 41.89303], "name": "Curia Hostilia (Senate House of Hostilius)", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.484446465969086, 41.89311723143024], "name": "Carcer (Ancus's Prison)", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.307863235473633, 41.76695862392542], "name": "Ostia", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.640371322631836, 41.73263596155049], "name": "Apiolae", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.484999001026154, 41.89290159672141], "name": "Comitium (Assembly Place)", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.666560411453247, 41.92475971933975], "name": "Collatia", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.481812536716461, 41.89230660087878], "name": "Temple to Jupiter Optimus Maximus", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.4873486161232, 41.890976825034336], "name": "Palace of Tarquinius Priscus", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.482064664363861, 41.88272999936073], "name": "Temple to Diana", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.490223944187164, 41.892450358771384], "name": "Cuprius Street", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.705988883972168, 41.80151886731471], "name": "Tusculum", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.66397476196289, 41.71264864888721], "name": "Grove of Feretina", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.536612749099731, 41.861666810506186], "name": "Gabii", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.481681108474731, 41.8920390506062], "name": "Shrine to Terminus", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.480157613754272, 41.888748489937704], "name": "Cloaca Maxima (Great Sewer)", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [13.02103042602539, 41.69009139667866], "name": "Signia", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [13.044977188110352, 41.23767298741489], "name": "Ciceii", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.512054443359375, 41.60722821271717], "name": "Ardea", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.093887329101562, 41.98909812021334], "name": "Caere", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.485674917697906, 41.89247431838869], "name": "Shrine to Venus Cloacina", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.397513389587402, 42.02602503159178], "name": "Veii", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [12.479411959648132, 41.8893554993159], "name": "Pons Sublicius (Wooden Pile Bridge)", "type": "Point"}, "properties": {"marker-color": "#00A0B0", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}], "type": "FeatureCollection"}
-var interest_zones = {"features": [{"geometry": {"coordinates": [[[12.709465026855467, 41.82454867985508], [12.798728942871094, 41.8148259912181], [12.861900329589844, 41.75184866809371], [12.79529571533203, 41.69650051162983], [12.699851989746094, 41.6872711837914], [12.610588073730469, 41.75492216766298], [12.565956115722656, 41.81329069491423], [12.595481872558592, 41.84859319874276], [12.709465026855467, 41.82454867985508]]], "name": "Alban Hills", "type": "Polygon"}, "properties": {"marker-color": "#8A9B0F", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [[[12.486766576766968, 41.886544038951655], [12.488311529159546, 41.88620857235047], [12.489620447158813, 41.8872948387714], [12.489513158798216, 41.88917978666802], [12.489786744117737, 41.88994652922092], [12.488134503364563, 41.890529566877404], [12.486385703086851, 41.89104870551747], [12.485495209693909, 41.890561513992466], [12.484561800956726, 41.889898608081005], [12.484003901481627, 41.88965900184206], [12.483757138252257, 41.888692580889085], [12.484014630317688, 41.888484918280156], [12.484508156776428, 41.88833316440825], [12.48490512371063, 41.88801368139437], [12.48496949672699, 41.88775809383291], [12.48515188694, 41.88759036393996], [12.486766576766968, 41.886544038951655]]], "name": "Palatine Hill", "type": "Polygon"}, "properties": {"marker-color": "#8A9B0F", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [[[12.48070478439331, 41.886879503791356], [12.481520175933838, 41.88716704368053], [12.482893466949461, 41.88646416611116], [12.485897541046143, 41.884355487005145], [12.485339641571045, 41.88266210339971], [12.484138011932373, 41.88100062681102], [12.479503154754639, 41.87975449101842], [12.478687763214111, 41.87981839600917], [12.476756572723389, 41.88278990750172], [12.48070478439331, 41.886879503791356]]], "name": "Aventine Hill", "type": "Polygon"}, "properties": {"marker-color": "#8A9B0F", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [[[12.483346760272978, 41.893368804337136], [12.483529150485992, 41.89333685862611], [12.483663260936737, 41.8932490078384], [12.483153641223907, 41.89278579259582], [12.483008801937103, 41.892873644020646], [12.482933700084685, 41.892985454750225], [12.483346760272978, 41.893368804337136]]], "name": "Grove of Asylum", "type": "Polygon"}, "properties": {"marker-color": "#8A9B0F", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [[[12.48222291469574, 41.894275257228834], [12.482276558876036, 41.893812049428874], [12.483006119728088, 41.89355648506678], [12.483778595924377, 41.89355648506678], [12.484293580055235, 41.8940755991063], [12.484207749366758, 41.89465061280956], [12.483531832695007, 41.89521763514195], [12.482973933219908, 41.895249579912274], [12.482255101203918, 41.895057911050564], [12.48222291469574, 41.894275257228834]]], "name": "Arx Capitolina (Citadel)", "type": "Polygon"}, "properties": {"marker-color": "#8A9B0F", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [[[12.4871689081192, 41.89109662579462], [12.487458586692808, 41.89152790667126], [12.484862208366394, 41.89267797477293], [12.484100461006165, 41.89205502378826], [12.487061619758606, 41.890984811758685], [12.4871689081192, 41.89109662579462]]], "name": "Forum Romanum", "type": "Polygon"}, "properties": {"marker-color": "#8A9B0F", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [[[12.477000653743744, 41.89913474345595], [12.477499544620514, 41.89894308625411], [12.477531731128693, 41.898428004673974], [12.477488815784454, 41.898088607563736], [12.476721704006195, 41.89805666421344], [12.47610479593277, 41.89816447295668], [12.476040422916412, 41.898643620724705], [12.47610479593277, 41.898967043435796], [12.47633010149002, 41.89908283635393], [12.477000653743744, 41.89913474345595]]], "name": "The Goat Swamp on the Campus Martius", "type": "Polygon"}, "properties": {"marker-color": "#8A9B0F", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [[[12.49147653579712, 41.88790186196206], [12.492120265960693, 41.88815744894835], [12.49319314956665, 41.88822134553516], [12.493922710418701, 41.88812550063099], [12.495038509368896, 41.88611272442658], [12.496111392974854, 41.8860168763589], [12.497742176055908, 41.88576128080878], [12.497870922088623, 41.88301356406514], [12.495768070220947, 41.88122428963741], [12.494995594024658, 41.881288193158284], [12.494094371795654, 41.88195917626922], [12.490060329437256, 41.88499448802313], [12.4904465675354, 41.886528064391555], [12.49121904373169, 41.886879503791356], [12.49147653579712, 41.88790186196206]]], "name": "Caelian Hill", "type": "Polygon"}, "properties": {"marker-color": "#8A9B0F", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [[[12.483440637588501, 41.88802965558301], [12.482239007949829, 41.88700729945743], [12.485178709030151, 41.885042412841855], [12.487946748733519, 41.88608077508668], [12.483440637588501, 41.88802965558301]]], "name": "Murciae (Murcian District)", "type": "Polygon"}, "properties": {"marker-color": "#8A9B0F", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [[[12.464032173156738, 41.89205502378826], [12.466521263122559, 41.888540827510525], [12.467207908630371, 41.88406793446202], [12.466435432434082, 41.880169872316046], [12.462830543518065, 41.87742191511894], [12.458367347717285, 41.8778053581006], [12.457165718078613, 41.884131835138994], [12.457680702209473, 41.89084105056772], [12.457680702209473, 41.893205082399206], [12.459053993225096, 41.89358843066795], [12.461714744567871, 41.89358843066795], [12.464032173156738, 41.89205502378826]]], "name": "Janiculum Hill", "type": "Polygon"}, "properties": {"marker-color": "#8A9B0F", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [[[12.483279705047607, 41.88790186196206], [12.482593059539795, 41.88723094125791], [12.483623027801514, 41.88640026776683], [12.484824657440184, 41.88592102814744], [12.486541271209717, 41.88480278838877], [12.487828731536863, 41.883940132916344], [12.488858699798584, 41.88483473836777], [12.483279705047607, 41.88790186196206]]], "name": "Circus Maximus", "type": "Polygon"}, "properties": {"marker-color": "#8A9B0F", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [[[12.483108043670654, 41.9020534518419], [12.486884593963623, 41.899721690058364], [12.490918636322021, 41.8999133449234], [12.493536472320555, 41.901957627692916], [12.494480609893799, 41.904768409638564], [12.49396562576294, 41.90684447594146], [12.489931583404541, 41.90738743599137], [12.48345136642456, 41.90540720646051], [12.483108043670654, 41.9020534518419]]], "name": "Quirinal Hill", "type": "Polygon"}, "properties": {"marker-color": "#8A9B0F", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [[[12.492291927337646, 41.89917866481704], [12.492291927337646, 41.89822037371484], [12.494051456451416, 41.89754956138713], [12.499845027923584, 41.89831620347219], [12.504651546478271, 41.901095203880594], [12.505896091461182, 41.903746321429416], [12.502763271331787, 41.904704529604835], [12.498171329498291, 41.90358661866875], [12.492291927337646, 41.89917866481704]]], "name": "Viminal Hill", "type": "Polygon"}, "properties": {"marker-color": "#8A9B0F", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}, {"geometry": {"coordinates": [[[12.515981197357176, 41.89093689139765], [12.516882419586182, 41.89208697014036], [12.513277530670166, 41.89480235167352], [12.50769853591919, 41.896144026994854], [12.49696969985962, 41.89550513749981], [12.492592334747313, 41.894323175085376], [12.491734027862549, 41.89275783984439], [12.49267816543579, 41.89138414670259], [12.496068477630615, 41.890361844261186], [12.503106594085693, 41.88965900184206], [12.515981197357176, 41.89093689139765]]], "name": "Esquiline Hill", "type": "Polygon"}, "properties": {"marker-color": "#8A9B0F", "marker-size": "medium", "marker-symbol": ""}, "type": "Feature"}], "type": "FeatureCollection"}
-var interest_lines = {"features": [], "type": "FeatureCollection"}
